+++ tfjs-core/src/packages/tfjs-backend-cpu/src/kernels/NonMaxSuppressionV5.ts
@@ -15,32 +15,20 @@
  * =============================================================================
  */
 
-import {NamedAttrMap, NamedTensorInfoMap, registerKernel, TensorInfo} from '../../kernel_registry';
-import {TypedArray} from '../../types';
-import {nonMaxSuppressionV5} from '../non_max_suppression_impl';
+import {NonMaxSuppressionV5, NonMaxSuppressionV5Attrs, NonMaxSuppressionV5Inputs} from '../../../../kernel_names';
+import {KernelConfig} from '../../../../kernel_registry';
+import {TypedArray} from '../../../../types';
+import {MathBackendCPU} from '../backend_cpu';
+import {nonMaxSuppressionV5Impl} from '../utils/non_max_suppression_impl';
+import {assertNotComplex} from '../utils/tensor_utils';
 
-import {MathBackendCPU} from './backend_cpu';
-import {assertNotComplex} from './cpu_util';
-
-interface NonMaxSuppressionWithScoreInputs extends NamedTensorInfoMap {
-  boxes: TensorInfo;
-  scores: TensorInfo;
-}
-
-interface NonMaxSuppressionWithScoreAttrs extends NamedAttrMap {
-  maxOutputSize: number;
-  iouThreshold: number;
-  scoreThreshold: number;
-  softNmsSigma: number;
-}
-
-registerKernel({
-  kernelName: 'NonMaxSuppressionV5',
+export const nonMaxSuppressionV5Config: KernelConfig = {
+  kernelName: NonMaxSuppressionV5,
   backendName: 'cpu',
   kernelFunc: ({inputs, backend, attrs}) => {
-    const {boxes, scores} = inputs as NonMaxSuppressionWithScoreInputs;
+    const {boxes, scores} = inputs as NonMaxSuppressionV5Inputs;
     const {maxOutputSize, iouThreshold, scoreThreshold, softNmsSigma} =
-        attrs as NonMaxSuppressionWithScoreAttrs;
+        attrs as unknown as NonMaxSuppressionV5Attrs;
 
     const cpuBackend = backend as MathBackendCPU;
 
@@ -54,10 +42,10 @@ registerKernel({
     const scoreThresholdVal = scoreThreshold;
     const softNmsSigmaVal = softNmsSigma;
 
-    const {selectedIndices, selectedScores} = nonMaxSuppressionV5(
+    const {selectedIndices, selectedScores} = nonMaxSuppressionV5Impl(
         boxesVals, scoresVals, maxOutputSizeVal, iouThresholdVal,
         scoreThresholdVal, softNmsSigmaVal);
 
     return [selectedIndices, selectedScores];
   }
-});
+};
