+++ tfjs-core/src/packages/tfjs-backend-cpu/src/utils/kernel_utils.ts
@@ -16,10 +16,10 @@
  */
 
 // TODO(yassogba) Import utils directly until we can break the cyclic dependency
-import * as backend_util from '../../../backends/backend_util';
+import * as backend_util from '../../../../backends/backend_util';
 // TODO(yassogba) export types from core.
-import {DataType, NumericDataType, TypedArray} from '../../../types';
-import * as util from '../../../util';
+import {DataType, NumericDataType, TypedArray} from '../../../../types';
+import * as util from '../../../../util';
 
 export function broadcastedBinaryOp(
     aShape: number[], bShape: number[], aVals: TypedArray, bVals: TypedArray,
